fix: correct obsidian's syntax tree.
Currently for the string
$|$
... text
$$
...
$$
where | is the cursor, is the first part recognized as blockmath instead of the second part. Essentially the cursor is outside the math environment but we treat it as if its inside. This then leads to the error behaviour in the autofraction and while technically correct by obsidians syntax, laggy behaviour. This also fixes the inconsistency of the math-preview being on if you're at top $|$ but not being on at the bottom $|$. Now it's never on when you are between $|$.
Fixes #415 #407