add matrix enter shortcut to vim
give the option to redefine the `insert new line below` to also append a
\\ at end of the current line in a matrix environment. The enter is
handled by simulating a enter press call. The cursor is moved to the
current line and collum = line.length + 1. This may have unwanted
consequences, other option would be to handle the key with vim but that
relies on the vim keymap.